+++ read_structure_step/formats/cif/cif.py
@@ -62,7 +62,7 @@ def load_cif(
     add_hydrogens=False,
     system_db=None,
     system=None,
-    indices="1-end",
+    indices="1:end",
     subsequent_as_configurations=False,
     system_name="from file",
     configuration_name="sequential",
@@ -96,7 +96,7 @@ def load_cif(
     system : System = None
         The system to use if adding subsequent structures as configurations.
 
-    indices : str = "1-end"
+    indices : str = "1:end"
         The generalized indices (slices, SMARTS, etc.) to select structures
         from a file containing multiple structures.