+++ read_structure_step/formats/extxyz/extxyz.py
@@ -68,7 +68,7 @@ def load_extxyz(
     add_hydrogens=False,
     system_db=None,
     system=None,
-    indices="1-end",
+    indices="1:end",
     subsequent_as_configurations=False,
     system_name="from file",
     configuration_name="sequential",
@@ -105,7 +105,7 @@ def load_extxyz(
     system : System = None
         The system to use if adding subsequent structures as configurations.
 
-    indices : str = "1-end"
+    indices : str = "1:end"
         The generalized indices (slices, SMARTS, etc.) to select structures
         from a file containing multiple structures.